+++ example/lib/main.dart
@@ -23,6 +23,7 @@ void main() async {
     // startLocale: Locale('de', 'DE'),
     // saveLocale: false,
     // useOnlyLangCode: true,
+    // ignorePluralRules: false,
 
     // optional assetLoader default used is RootBundleAssetLoader which uses flutter's assetloader
     // install easy_localization_loader for enable custom loaders