+++ lib/src/easy_localization_app.dart
@@ -54,6 +54,35 @@ class EasyLocalization extends StatefulWidget {
   /// ```
   final bool useFallbackTranslations;
 
+  /// If a localization key is empty in the locale file, try to use the fallbackLocale file.
+  /// Does not take effect if [useFallbackTranslations] is false.
+  /// @Default value false
+  /// Example:
+  /// ```
+  /// useFallbackTranslationsForEmptyResources: true
+  /// ```
+  final bool useFallbackTranslationsForEmptyResources;
+
+  /// Ignore usage of plural strings for languages that do not use plural rules.
+  /// @Default value false
+  /// Example:
+  /// ```
+  /// // Default behavior, use "zero" rule for 0 even if the language doesn't
+  /// // use it by default (e.g. "en"). If "zero" localization for that string
+  /// // doesn't exist, "other" is still used as fallback.
+  /// // "nTimes": "{count, plural, =0{never} =1{once} other{{count} times}}"
+  /// // Text(AppLocalizations.of(context)!.nTimes(_counter)),
+  /// // will print "never, once, 2 times" for ALL languages.
+  /// ignorePluralRules: true
+  /// // Use "zero" rule for 0 only if the language is set to do so (e.g. for
+  /// "lt" but not for "en").
+  /// // "nTimes": "{count, plural, =0{never} =1{once} other{{count} times}}"
+  /// // Text(AppLocalizations.of(context)!.nTimes(_counter)),
+  /// // will print "never, once, 2 times" ONLY for languages with plural rules.
+  /// ignorePluralRules: false
+  /// ```
+  final bool ignorePluralRules;
+
   /// Path to your folder with localization files.
   /// Example:
   /// ```dart
@@ -65,7 +94,30 @@ class EasyLocalization extends StatefulWidget {
   /// Class loader for localization files.
   /// You can use custom loaders from [Easy Localization Loader](https://github.com/aissat/easy_localization_loader) or create your own class.
   /// @Default value `const RootBundleAssetLoader()`
-  final assetLoader;
+  // ignore: prefer_typing_uninitialized_variables
+  final AssetLoader assetLoader;
+
+  /// Class loader for localization files that belong to other packages.
+  /// You can use custom loaders from [Easy Localization Loader](https://github.com/aissat/easy_localization_loader) or create your own class.
+  /// Example:
+  /// ```dart
+  //   runApp(
+  //   EasyLocalization(
+  //     supportedLocales: const <Locale>[
+  //       Locale('en'),
+  //     ],
+  //     fallbackLocale: const Locale('en'),
+  //     assetLoader: const RootBundleAssetLoader(),
+  //     extraAssetLoaders: [
+  //         TranslationsLoader(packageName: 'package_example_1'),
+  //         TranslationsLoader(packageName: 'package_example_2'),
+  //     ],
+  //     path: 'lib/l10n/translations',
+  //     child: const MainApp(),
+  //   ),
+  // );
+  /// @Default value `null`
+  final List<AssetLoader>? extraAssetLoaders;
 
   /// Save locale in device storage.
   /// @Default value true
@@ -84,7 +136,10 @@ class EasyLocalization extends StatefulWidget {
     this.startLocale,
     this.useOnlyLangCode = false,
     this.useFallbackTranslations = false,
+    this.useFallbackTranslationsForEmptyResources = false,
+    this.ignorePluralRules = true,
     this.assetLoader = const RootBundleAssetLoader(),
+    this.extraAssetLoaders,
     this.saveLocale = true,
     this.errorWidget,
   })  : assert(supportedLocales.isNotEmpty),
@@ -94,8 +149,10 @@ class EasyLocalization extends StatefulWidget {
   }
 
   @override
+  // ignore: library_private_types_in_public_api
   _EasyLocalizationState createState() => _EasyLocalizationState();
 
+  // ignore: library_private_types_in_public_api
   static _EasyLocalizationProvider? of(BuildContext context) =>
       _EasyLocalizationProvider.of(context);
 
@@ -123,6 +180,7 @@ class _EasyLocalizationState extends State<EasyLocalization> {
       supportedLocales: widget.supportedLocales,
       startLocale: widget.startLocale,
       assetLoader: widget.assetLoader,
+      extraAssetLoaders: widget.extraAssetLoaders,
       useOnlyLangCode: widget.useOnlyLangCode,
       useFallbackTranslations: widget.useFallbackTranslations,
       path: widget.path,
@@ -159,6 +217,9 @@ class _EasyLocalizationState extends State<EasyLocalization> {
       delegate: _EasyLocalizationDelegate(
         localizationController: localizationController,
         supportedLocales: widget.supportedLocales,
+        useFallbackTranslationsForEmptyResources:
+            widget.useFallbackTranslationsForEmptyResources,
+        ignorePluralRules: widget.ignorePluralRules,
       ),
     );
   }
@@ -204,14 +265,15 @@ class _EasyLocalizationProvider extends InheritedWidget {
 
   /// Get fallback locale
   Locale? get fallbackLocale => parent.fallbackLocale;
+
   // Locale get startLocale => parent.startLocale;
 
   /// Change app locale
-  Future<void> setLocale(Locale _locale) async {
+  Future<void> setLocale(Locale locale) async {
     // Check old locale
-    if (_locale != _localeState.locale) {
-      assert(parent.supportedLocales.contains(_locale));
-      await _localeState.setLocale(_locale);
+    if (locale != _localeState.locale) {
+      assert(parent.supportedLocales.contains(locale));
+      await _localeState.setLocale(locale);
     }
   }
 
@@ -222,6 +284,7 @@ class _EasyLocalizationProvider extends InheritedWidget {
 
   /// Getting device locale from platform
   Locale get deviceLocale => _localeState.deviceLocale;
+  Locale? get savedLocale => _localeState.savedLocale;
 
   /// Reset locale to platform locale
   Future<void> resetLocale() => _localeState.resetLocale();
@@ -238,12 +301,18 @@ class _EasyLocalizationProvider extends InheritedWidget {
 class _EasyLocalizationDelegate extends LocalizationsDelegate<Localization> {
   final List<Locale>? supportedLocales;
   final EasyLocalizationController? localizationController;
+  final bool useFallbackTranslationsForEmptyResources;
+  final bool ignorePluralRules;
 
   ///  * use only the lang code to generate i18n file path like en.json or ar.json
   // final bool useOnlyLangCode;
 
-  _EasyLocalizationDelegate(
-      {this.localizationController, this.supportedLocales}) {
+  _EasyLocalizationDelegate({
+    required this.useFallbackTranslationsForEmptyResources,
+    this.ignorePluralRules = true,
+    this.localizationController,
+    this.supportedLocales,
+  }) {
     EasyLocalization.logger.debug('Init Localization Delegate');
   }
 
@@ -257,9 +326,14 @@ class _EasyLocalizationDelegate extends LocalizationsDelegate<Localization> {
       await localizationController!.loadTranslations();
     }
 
-    Localization.load(value,
-        translations: localizationController!.translations,
-        fallbackTranslations: localizationController!.fallbackTranslations);
+    Localization.load(
+      value,
+      translations: localizationController!.translations,
+      fallbackTranslations: localizationController!.fallbackTranslations,
+      useFallbackTranslationsForEmptyResources:
+          useFallbackTranslationsForEmptyResources,
+      ignorePluralRules: ignorePluralRules,
+    );
     return Future.value(Localization.instance);
   }